script/pipeline_mongo/requirements.txt
pymongo>=4.6.0
rapidfuzz>=3.9.0
openai>=1.50.0
pymilvus>=2.4.6
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
